# docs/openapi.yaml
openapi: 3.1.0
info: {title: PPT Agent P8 API, version: 1.0.0-step1}
components:
  parameters:
    TaskId: {in: path, name: task_id, required: true, schema: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$'}}
    JobId: {in: path, name: job_id, required: true, schema: {type: string, pattern: '^job_[A-Za-z0-9_-]+$'}}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: object, required: [code, message, diagnostic_id], properties: {code: {type: string}, message: {type: string}, diagnostic_id: {type: string}, agent_audit_id: {type: string}}}
...
    AgentAudits: {type: object, required: [audits], properties: {audits: {type: array, items: {type: object}}}}
...
    JobCreate:
      type: object
      additionalProperties: false
      required: [operation, idempotency_key]
      properties:
        operation: {enum: [clarification.generate, narrative.generate, outline.generate, samples.generate, samples.modify, deck.generate, deck.modify, inspection.run]}
        payload: {type: object}
        idempotency_key: {type: string, minLength: 1, maxLength: 200}
    Job:
      type: object
      required: [job_id, task_id, operation, status, progress, current_step, last_seq, created_at, started_at, finished_at, result, error, cancellation_requested]
      properties:
        job_id: {type: string, pattern: '^job_[A-Za-z0-9_-]+$'}
        task_id: {type: string}
        operation: {type: string}
        status: {enum: [queued, running, cancellation_requested, succeeded, failed, cancelled, interrupted]}
        progress: {type: [integer, 'null'], minimum: 0, maximum: 100}
        current_step: {type: string}
        last_seq: {type: integer, minimum: 0}
        created_at: {type: string, format: date-time}
        started_at: {type: [string, 'null'], format: date-time}
        finished_at: {type: [string, 'null'], format: date-time}
        result: {type: [object, 'null']}
        error: {type: [object, 'null']}
        cancellation_requested: {type: boolean}
    Jobs: {type: object, required: [jobs], properties: {jobs: {type: array, items: {$ref: '#/components/schemas/Job'}}}}
  responses:
    Error: {description: Unified domain error, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}, example: {error: {code: validation_error, message: 请求字段无效, diagnostic_id: abc123}}}}}
paths:
...
  /v1/jobs/{job_id}:
    get:
      parameters: [{$ref: '#/components/parameters/JobId'}]
      responses: {'200': {description: Authoritative persistent Job snapshot, content: {application/json: {schema: {$ref: '#/components/schemas/Job'}, example: &job {job_id: job_abc123, task_id: demo, operation: deck.generate, status: queued, progress: 0, current_step: queued, last_seq: 1, created_at: '2026-08-13T00:00:00+00:00', started_at: null, finished_at: null, result: null, error: null, cancellation_requested: false}}}}, '404': {$ref: '#/components/responses/Error'}}
  /v1/jobs/{job_id}/agent-audits:
    get:
      parameters: [{$ref: '#/components/parameters/JobId'}]
      responses: {'200': {description: Secret-free Agent audits correlated to this Job, content: {application/json: {schema: {$ref: '#/components/schemas/AgentAudits'}, example: {audits: []}}}}, '404': {$ref: '#/components/responses/Error'}}
  /v1/jobs/{job_id}/cancel:
    post:
      parameters: [{$ref: '#/components/parameters/JobId'}]
      requestBody: {content: {application/json: {schema: {type: object, additionalProperties: false}}}}
      responses: {'200': {description: Cancellation requested or terminal snapshot returned, content: {application/json: {schema: {$ref: '#/components/schemas/Job'}, example: *job}}}, '404': {$ref: '#/components/responses/Error'}}
  /v1/jobs/{job_id}/events:
    get:
      parameters:
        - {$ref: '#/components/parameters/JobId'}
        - {in: query, name: after, schema: {type: integer, minimum: 0, default: 0}}
        - {in: header, name: Last-Event-ID, schema: {type: string}}
      responses:
        '200': {description: Ordered resumable Job events and heartbeat, content: {text/event-stream: {schema: {type: string}, example: "id: 1\nevent: queued\ndata: {\"seq\":1,\"type\":\"queued\"}\n\n"}}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
  /v1/tasks/{task_id}/jobs:
    post:
      parameters: [{$ref: '#/components/parameters/TaskId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/JobCreate'}}}}
      responses: {'202': {description: New or idempotently recovered Job, content: {application/json: {schema: {$ref: '#/components/schemas/Job'}, example: *job}}}, '400': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}
    get:
      parameters:
        - {$ref: '#/components/parameters/TaskId'}
        - {in: query, name: status, schema: {enum: [active, queued, running, cancellation_requested, succeeded, failed, cancelled, interrupted]}}
      responses: {'200': {description: Job discovery and history, content: {application/json: {schema: {$ref: '#/components/schemas/Jobs'}, example: {jobs: [*job]}}}}, '404': {$ref: '#/components/responses/Error'}}
...